+++ lib/scss_lint/linter/url_quotes.rb
@@ -19,6 +19,8 @@ module SCSSLint
 
     def check(node, string)
       return unless string =~ /^\s*url\(\s*[^"']/
+      return if string =~ /^\s*url\(\s*data:/ # Ignore data URIs
+
       add_lint(node, 'URLs should be enclosed in quotes')
     end
   end